Aula 33 dialog de confirmação se deseja remover mesmo o curso

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmationDialogComponent } from './../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
@@ -68,14 +69,22 @@ export class CoursesComponent implements OnInit {
     console.log('onEdit');
   }
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(() => {
-        this.refresh();
-        this.snackBar
-        .open('Curso removido com sucesso!', 'X',
-        { duration: 3000, verticalPosition: 'top', horizontalPosition:'center' });
-      },
-      error => this.onError('Erro ao tentar remover curso.')
-      );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(() => {
+          this.refresh();
+          this.snackBar
+          .open('Curso removido com sucesso!', 'X',
+          { duration: 3000, verticalPosition: 'top', horizontalPosition:'center' });
+        },
+        error => this.onError('Erro ao tentar remover curso.')
+        );
+      }
+    });
   }
 
 }
